Better deal with test tips
