+++ packages/tests/src/_jest/globalSetup.ts
@@ -1,0 +1,8 @@
+import { logTips } from './helpers/tips';
+
+const globalSetup = () => {
+    // Log some tips to the console.
+    logTips();
+};
+
+export default globalSetup;
